chore(core): deduplicate retry logic, tighten module visibility, clean up logs
- Extract shared with_retry() into sonos/retry.rs
- Narrow grouping, playback, volume, zone_groups to pub(crate)
- Downgrade DIDL diagnostic logs from info to debug
- Document intentional retry omission on stop/switch_to_queue
